chore : CaptureInfo 타입 변경
