--- .prettierrc.cjs
@@ -1,9 +1,0 @@
-module.exports = {
-  endOfLine: 'auto',
-  printWidth: 80,
-  semi: true,
-  singleQuote: true,
-  tabWidth: 2,
-  trailingComma: 'all',
-  useTabs: false,
-};
