+++ package.json
@@ -44,7 +44,7 @@
   "private": true,
   "scripts": {
     "build": "next build",
-    "dev": "next dev -p 3004",
+    "dev": "next dev -p 3006",
     "lint": "next lint",
     "start": "next start"
   },
