+++ utils/fcm.js
@@ -28,7 +28,7 @@ getToken(messaging, {
     if (currentToken) {
       // Send the token to your server and update the UI if necessary
       // ...
-      console.log(currentToken);
+      console.log('currentToken', currentToken);
     } else {
       // Show permission request UI
       console.log(
